+++ simple-stack/src/main/java/com/zhuinden/simplestack/ServiceSearchMode.java
@@ -1,0 +1,9 @@
+package com.zhuinden.simplestack;
+
+/**
+ * {@link ServiceSearchMode} determines whether the services of the parent scopes should be included along with the found services.
+ */
+public enum ServiceSearchMode {
+    LOCAL_SERVICES_ONLY,
+    INCLUDE_PARENT_SERVICE;
+}