+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 -Simple Stack X.X.X (XXXX-XX-XX)
 --------------------------------
 
-- Added: `Backstack.goAppendChain(newKeys)` and `Backstack.goAppendChain(asReplace, newKeys)` which appends the provided keys to the end of
+- ADDED: `Backstack.goAppendChain(newKeys)` and `Backstack.goAppendChain(asReplace, newKeys)` which appends the provided keys to the end of
   the current history.
 
 If any duplicates are provided, they will also be added to the end. If a key already exists in the history, it gets removed from earlier and
@@ -11,6 +11,11 @@ appended to the end.
 
 If there are implicit parents used as the reorder occurs, please make sure that the scope hierarchy can still be rebuilt from left-to-right
 order. It might be preferred top use `ScopeKey.Child` instead of `ScopeKey` in these cases.
+
+- ADDED: `Backstack.findServices(serviceSearchMode)` and `Backstack.findServicesFromScope(scopeTag, serviceSearchMode)`.
+
+This allows for getting the services of a backstack (either only local services, or including parent services) that are accessible within
+the backstack.
 
 -Simple Stack 2.8.0 (2023-07-03)
 --------------------------------